fix: the architect re-review's minors — gate needle, one-copy ruling, guard rationale (#677, #618)
SizeBoundGateNeedleTests pins the observe gate's production
default (?? didRecentlySetFrame) — the branch every fixture
severs — and the balanced-body walker the needles share is
extracted to SourceScan.functionBody at its third consumer
(ZOrderSequenceWiringTests predates the extraction and keeps
its private copy). The tab-rekey acceptance now lives in ONE
copy, at rekeyCandidates, with the coalescer header reduced to
the pointer. The delivery guard states why observer presence,
not the full ownsObservation funnel, is the right narrowing
for a read-flight interval (a policy flip still ends in a
detach; the funnel's extra term costs an NSRunningApplication
lookup per delivered frame at storm rate).
