+++ Tests/KiwiDeskGuiTests/OwnPidQueueNeedleTests.swift
@@ -18,60 +18,9 @@ import Testing
 /// argument lives on the arms themselves.
 @Suite("Own-pid queue needles (#618)")
 struct OwnPidQueueNeedleTests {
-    private func body(
-        of function: String,
-        in file: String,
-        under directory: String,
-        _ path: StaticString = #filePath
-    ) throws -> String {
-        let url = SourceScan.repoRoot(from: "\(path)")
-            .appendingPathComponent("Sources/KiwiDeskCore")
-            .appendingPathComponent(directory)
-            .appendingPathComponent(file)
-        let source = SourceScan.stripComments(
-            try String(contentsOf: url, encoding: .utf8)
-        )
-        let characters = Array(source)
-        let marker = Array("func \(function)(")
-        guard
-            let start = (0...(characters.count - marker.count))
-                .first(where: { index in
-                    Array(
-                        characters[index..<(index + marker.count)]
-                    ) == marker
-                })
-        else {
-            Issue.record("\(function) not found in \(file)")
-            return ""
-        }
-        var cursor = start + marker.count - 1
-        guard
-            SourceScan.balanced(
-                characters,
-                from: &cursor,
-                open: "(",
-                close: ")"
-            ) != nil
-        else {
-            Issue.record("\(function): signature not balanced")
-            return ""
-        }
-        while cursor < characters.count,
-            characters[cursor] != "{"
-        {
-            cursor += 1
-        }
-        return SourceScan.balanced(
-            characters,
-            from: &cursor,
-            open: "{",
-            close: "}"
-        ) ?? ""
-    }
-
     @Test("The frame-read queue keeps its own-pid main arm")
     func frameReadQueueKeepsOwnPidArm() throws {
-        let source = try body(
+        let source = try SourceScan.functionBody(
             of: "queue",
             in: "FrameReadCoalescer.swift",
             under: "Events"
@@ -82,7 +31,7 @@ struct OwnPidQueueNeedleTests {
 
     @Test("The frame-apply queue keeps its own-pid main arm")
     func frameApplyQueueKeepsOwnPidArm() throws {
-        let source = try body(
+        let source = try SourceScan.functionBody(
             of: "queue",
             in: "FrameApplier.swift",
             under: "Tiling"
